+++ flow-typed/type.js
@@ -1,0 +1,19 @@
+// @flow
+'use strict'
+
+declare type UserInformation = {
+  userId: string,
+  email: string | void,
+  firstName: string | void,
+  lastName: string | void,
+  fullName: string | void,
+  picture: string | void,
+  role: string | void,
+  username: string,
+  providerType: string,
+  providerUserId: string,
+  isSAMLUser: boolean,
+  supervisorFullName: string | null,
+  supervisorEmail: string | null,
+  supervisorProviderUserId: string | null,
+}